simplify migrations
simplified migrations and change order of mirror transfers to start with well headers
